moved tests to node_test.go. Delete setup_test.go file
